refactor: extract PermissionGroupAccordion into it's own component
CU-86c8ea3m5
